+++ packages/common/src/core/schemas/internal/metrics/StatCardUiSchema.ts
@@ -59,6 +59,7 @@ export const buildUiSchema = (
             scope: "/properties/allowUnitFormatting",
             options: {
               section: "subblock",
+              scale: "m",
               toggleDisplay: "switch",
               scope: "/properties/unit",
             },
@@ -94,6 +95,7 @@ export const buildUiSchema = (
             labelKey: `advancedConfig.label`,
             options: {
               section: "subblock",
+              scale: "m",
             },
             rule: SHOW_FOR_DYNAMIC_RULE,
             elements: [
@@ -163,6 +165,7 @@ export const buildUiSchema = (
             scope: "/properties/allowLink",
             options: {
               section: "subblock",
+              scale: "m",
               toggleDisplay: "switch",
             },
             rule: SHOW_FOR_STATIC_RULE,