+++ packages/common/src/projects/_internal/ProjectUiSchemaCreate.ts
@@ -25,8 +25,11 @@ export const buildUiSchema = async (
         options: { section: "stepper", scale: "l" },
         elements: [
           {
-            type: "Step",
+            type: "Section",
             labelKey: `${i18nScope}.sections.details.label`,
+            options: {
+              section: "step",
+            },
             elements: [
               {
                 type: "Section",
@@ -90,8 +93,11 @@ export const buildUiSchema = async (
             ],
           },
           {
-            type: "Step",
+            type: "Section",
             labelKey: `${i18nScope}.sections.location.label`,
+            options: {
+              section: "step",
+            },
             rule: {
               effect: UiSchemaRuleEffects.DISABLE,
               condition: {
@@ -132,8 +138,11 @@ export const buildUiSchema = async (
             ],
           },
           {
-            type: "Step",
+            type: "Section",
             labelKey: `${i18nScope}.sections.sharing.label`,
+            options: {
+              section: "step",
+            },
             rule: {
               effect: UiSchemaRuleEffects.DISABLE,
               condition: {