+++ packages/common/src/templates/_internal/computeProps.ts
@@ -7,6 +7,7 @@ import { IHubTemplate } from "../../core/types/IHubTemplate";
 import { computeLinks } from "./computeLinks";
 import { getProp } from "../../objects";
 import { getDeployedTemplateType } from "../utils";
+import { computeBaseProps } from "../../core/_internal/computeBaseProps";
 
 /**
  * @private
@@ -22,33 +23,36 @@ export function computeProps(
   template: Partial<IHubTemplate>,
   requestOptions: IRequestOptions
 ): IHubTemplate {
-  // 1. compute relevant template links
+  // 1. compute base properties on template
+  template = computeBaseProps(model.item, template);
+
+  // 2. compute relevant template links
   template.links = computeLinks(model.item, requestOptions);
 
-  // 2. append the template's thumbnail url at the top-level
+  // 3. append the template's thumbnail url at the top-level
   template.thumbnailUrl = template.links.thumbnail;
 
-  // 3. compute relevant template dates
+  // 4. compute relevant template dates
   template.createdDate = new Date(model.item.created);
   template.createdDateSource = "item.created";
   template.updatedDate = new Date(model.item.modified);
   template.updatedDateSource = "item.modified";
 
-  // 4. determine whether the template is discussable
+  // 5. determine whether the template is discussable
   template.isDiscussable = isDiscussable(template);
 
-  // 5. process features that can be disabled by the entity owner
+  // 6. process features that can be disabled by the entity owner
   template.features = processEntityFeatures(
     getProp(model, "data.settings.features") || {},
     TemplateDefaultFeatures
   );
 
-  // 6. compute additional template-specific properties
+  // 7. compute additional template-specific properties
   template.isDeployed = (getProp(model, "item.typeKeywords") || []).includes(
     "Deployed"
   );
   template.deployedType = getDeployedTemplateType(model.item);
 
-  // 7. cast b/c this takes a partial but returns a full template
+  // 8. cast b/c this takes a partial but returns a full template
   return template as IHubTemplate;
 }